chore(Docker): Don't run tests when building in Docker
Since the integration tests try to spin up a Docker container, doing so
within a Docker build would be weird. I'm sure there's some way to do
it, but I'm not sure how and I'm not sure if it's a good idea. Since
integ tests run in CI, it seems OK to remove them from the Docker build.